+++ packages/jest-validate/src/__tests__/validate_cli_options.test.js
@@ -8,7 +8,7 @@
 
 'use strict';
 
-import validateCLIOptions from '../validate_cli_options';
+import validateCLIOptions from '../validateCLIOptions';
 
 test('validates yargs special options', () => {
   const options = ['$0', '_', 'help', 'h'];
