+++ packages/jest-validate/src/exampleConfig.js
