+++ packages/jest-validate/src/index.js
@@ -14,7 +14,7 @@ import {
   ValidationError,
 } from './utils';
 import validate from './validate';
-import validateCLIOptions from './validate_cli_options';
+import validateCLIOptions from './validateCLIOptions';
 import {multipleValidOptions} from './condition';
 
 module.exports = {
